+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add view button to update prompt
 - Fix context menu open in new tab 
 - Fix context menu open in new window
+- Bookmark shortcuts work even when webview is focused
 
 ## v0.13.0 - 2026-03-21